# GS: Check whole state before flushing draws