--- a/pcsx2/GS/GSState.h
+++ b/pcsx2/GS/GSState.h
@@ -235,8 +235,8 @@ public:
 	int m_frameskip;
 	bool m_nativeres;
 	bool m_mipmap;
-	bool m_reg_texture_dirty;
-	bool m_reg_draw_dirty;
+	u32 m_dirty_gs_regs;
+	int m_backed_up_ctx;
 
 	static int s_n;
 	bool s_dump;
@@ -249,6 +249,30 @@ public:
 	std::string m_dump_root;
 
 	static constexpr u32 STATE_VERSION = 8;
+
+	enum REG_DIRTY
+	{
+		DIRTY_REG_ALPHA,
+		DIRTY_REG_CLAMP,
+		DIRTY_REG_COLCLAMP,
+		DIRTY_REG_DIMX,
+		DIRTY_REG_DTHE,
+		DIRTY_REG_FBA,
+		DIRTY_REG_FOGCOL,
+		DIRTY_REG_FRAME,
+		DIRTY_REG_MIPTBP1,
+		DIRTY_REG_MIPTBP2,
+		DIRTY_REG_PABE,
+		DIRTY_REG_PRIM,
+		DIRTY_REG_SCANMSK,
+		DIRTY_REG_SCISSOR,
+		DIRTY_REG_TEST,
+		DIRTY_REG_TEX0,
+		DIRTY_REG_TEX1,
+		DIRTY_REG_TEXA,
+		DIRTY_REG_XYOFFSET,
+		DIRTY_REG_ZBUF
+	};
 
 	enum PRIM_OVERLAP
 	{
@@ -317,8 +341,9 @@ public:
 	virtual void Reset(bool hardware_reset);
 	virtual void UpdateSettings(const Pcsx2Config::GSOptions& old_config);
 
-	void Flush(bool forced = false);
-	void FlushPrim(bool forced);
+	void CopyEnv(GSDrawingEnvironment* dest, GSDrawingEnvironment* src, int ctx);
+	void Flush();
+	void FlushPrim();
 	bool TestDrawChanged();
 	void FlushWrite();
 	virtual void Draw() = 0;