--- a/pcsx2/GS/Renderers/Common/GSRenderer.cpp
+++ b/pcsx2/GS/Renderers/Common/GSRenderer.cpp
@@ -504,7 +504,7 @@ static GSVector4 CalculateDrawRect(s32 window_width, s32 window_height, s32 text
 
 void GSRenderer::VSync(u32 field, bool registers_written)
 {
-	Flush(true);
+	Flush();
 
 	if (s_dump && s_n >= s_saven)
 	{