### Description of Changes
Changes the GS State to check the entire register state on vertex request before flushing draws.

### Rationale behind Changes
Before we were super paranoid with flushing, but some games trash the register states then put them back as they were, before this would have caused it to flush, which could reduce the framerate and inflate the draw call count (not so good for OGL or weak GPUs)

### Suggested Testing Steps
Just test games, make sure now graphics are broken, record any differences in performance Vs master.


Here's some of my results. NOTE: All my testing was done with GS Dumps at x3 Native in Vulkan on a 2080Ti + Ryzen 5900x, this may not translate completely to playing the game!

D = Draw Calls
TC  = Texture Copies

FPS gains observed:
-------------------------------------------------------------------------------
Driv3r:  35% increase.
Master: 1643D, 137fps, 378TC
PR: 1422D, 185fps, 17TC
(with autoflush)
Master: 1670D, 134fps, 390TC
PR: 1450D, 180fps, 29TC

Genji: 17% increase.
Master: 423D, 790fps
PR: 169D, 930fps

GTA Vice City: 1% increase.
Master: 271D, 955fps
PR: 269D, 963fps

MGS2 Thermal Vision: 10% increase.
Master: 717D, 427fps
PR: 551D, 470fps

Parappa The Rapper 2: (Bonus it fixed a mipmapping bug) 17% increase.
Master: 390-402D, 1296fps
PR: 335-343D, 1518fps